+++ packages/adhere-util/src/url.ts
@@ -77,6 +77,62 @@ const methods = {
 
     return `?${getStr.join('&')}`;
   },
+
+  /**
+   * getPathName
+   * @description 不同路由模式下获取pathname的方法
+   * @return {string}
+   */
+  getPathName(publicPath: string = '/', router: 'hash' | 'browser' = 'browser') {
+    const routerMode = router || 'browser';
+
+    let pathname = '';
+
+    if (routerMode === 'browser') {
+      pathname = window.location.pathname;
+    } else if (routerMode === 'hash') {
+      const hash = window.location.hash;
+      if (hash.lastIndexOf('?') !== -1) {
+        pathname = hash.substring(1, hash.lastIndexOf('?'));
+      } else {
+        pathname = hash.substring(1);
+      }
+    }
+
+    if (publicPath !== '/') {
+      pathname = pathname.replace(`${publicPath}/`, '');
+    }
+
+    return pathname;
+  },
+  /**
+   * getSearch
+   * @description 不同路由模式下获取search的方法
+   * @return {string}
+   */
+  getSearch(router: 'hash' | 'browser' = 'browser') {
+    const routerMode = router || 'browser';
+
+    if (routerMode === 'browser') {
+      return window.location.search;
+    } else if (routerMode === 'hash') {
+      const hash = window.location.hash;
+
+      const index = hash.lastIndexOf('?');
+      if (index !== -1) {
+        return hash.substring(index);
+      }
+
+      return '';
+    }
+  },
+  /**
+   * getFullPath
+   * @return {`${string}${string}`}
+   */
+  getFullPath() {
+    return `${methods.getPathName()}${methods.getSearch()}`;
+  },
 };
 
 // const memoizedMethods = {};